Move batch functions to separated file
