+++ src/cycleProcessing/batch.cpp
@@ -1,0 +1,226 @@
+#include "chrono"
+#include "atomic"
+#include "thread"
+
+#include "../IOmisc.h"
+#include "../fastExtractor.h"
+#include "../featureMatching.h"
+
+#include "mainCycleInternals.h"
+
+#include "batch.h"
+
+/**
+ * Fills batch up to dataProcessingConditions.batchSize or while new frames can be obtained.
+ * <br>
+ * Only frames with extracted points count greater or equal to
+ * dataProcessingConditions.requiredExtractedPointsCount will be added to batch
+ *
+ * @param [in] mediaInputStruct
+ * @param [in] dataProcessingConditions
+ * @param [in,out] currentBatch batch can contains some frames which was considered as bad for previous frame
+ *     but can be good for new one
+ *
+ * @return skipped frames count
+ */
+static int fillVideoFrameBatch(
+	MediaSources &mediaInputStruct,
+	const DataProcessingConditions &dataProcessingConditions,
+	std::vector<BatchElement> &currentBatch
+);
+
+/**
+ * Find new frame from batch which can be up to dataProcessingConditions.batchSize
+ *
+ * @param [in] mediaInputStruct
+ * @param [in] dataProcessingConditions
+ * @param [in,out] currentBatch <ul>
+ * 		<li>Input is a batch tail from previous iteration</li>
+ * 		<li>
+ * 			Output is a batch tail after new good frame
+ * 			(consequently there will be full batch in case we didn't find new good one)
+* 		</li>
+ * </ul>
+ * @param [in] previousFrame
+ * @param [out] newGoodFrame
+ * @param [in] previousFeatures
+ * @param [out] newFeatures
+ * @param [out] matches
+ * @return found frame batch index or some of error codes described above
+ */
+static int findGoodFrameFromBatch(
+	const DataProcessingConditions &dataProcessingConditions,
+	std::vector<BatchElement> &currentBatch,
+	std::vector<bool> &isMatchesEstimated,
+	std::vector<std::vector<DMatch>> &estimatedMatches,
+	Mat &previousFrame, Mat &newGoodFrame,
+	std::vector<KeyPoint> &previousFeatures,
+	std::vector<KeyPoint> &newFeatures,
+	std::vector<DMatch> &matches
+);
+
+int findGoodFrameFromBatchMultithreadingWrapper(
+	MediaSources &mediaInputStruct,
+	const DataProcessingConditions &dataProcessingConditions,
+	std::vector<BatchElement> &currentBatch,
+	Mat &previousFrame, Mat &newGoodFrame,
+	std::vector<KeyPoint> &previousFeatures,
+	std::vector<KeyPoint> &newFeatures,
+	std::vector<DMatch> &matches
+) {
+	fillVideoFrameBatch(mediaInputStruct, dataProcessingConditions, currentBatch);
+	int currentBatchSz = currentBatch.size();
+	if (currentBatchSz == 0)
+		return EMPTY_BATCH;
+
+	logStreams.mainReportStream << "Prev. extracted: " << previousFeatures.size() << std::endl;
+	auto start = std::chrono::high_resolution_clock::now();
+
+	std::vector<bool> isMatchesEstimated(currentBatchSz, false);
+	int threadsCnt = dataProcessingConditions.threadsCount;
+	std::vector<std::thread> threads;
+	std::vector<bool> threadShouldDieFlags(threadsCnt, false);
+	std::vector<std::vector<DMatch>> estimatedMatches(currentBatchSz, std::vector<DMatch>());
+	for (int i = 0; i < threadsCnt; ++i) {
+		threads.emplace_back(std::thread([&](int threadIndex) {
+			for (
+				int batchIndex = currentBatchSz - 1 - threadIndex;
+				batchIndex >= 0;
+				batchIndex -= threadsCnt
+			) {
+				matchFramesPairFeatures(previousFrame, currentBatch.at(batchIndex).frame,
+					previousFeatures, currentBatch.at(batchIndex).features,
+					dataProcessingConditions.matcherType, estimatedMatches.at(batchIndex));
+				isMatchesEstimated.at(batchIndex) = true;
+				if (threadShouldDieFlags.at(threadIndex))
+					break;
+			}
+		}, i));
+	}
+
+	int returnCode = findGoodFrameFromBatch(
+		dataProcessingConditions,
+		currentBatch, isMatchesEstimated, estimatedMatches,
+		previousFrame, newGoodFrame,
+		previousFeatures, newFeatures,
+		matches
+	);
+
+	std::cout << "Matching time for index " << returnCode << ": "
+			  << std::chrono::duration_cast<std::chrono::milliseconds>(
+				  std::chrono::high_resolution_clock::now() - start
+			  ).count() << std::endl;
+	for (int i = threadShouldDieFlags.size() - 1; i >= 0; --i)
+		threadShouldDieFlags.at(i) = true;
+	for (auto& thread : threads)
+		thread.join();
+	std::cout << "Threads terminated: "
+			  << std::chrono::duration_cast<std::chrono::milliseconds>(
+				  std::chrono::high_resolution_clock::now() - start
+			  ).count() << std::endl;
+
+	if (returnCode < 0)
+		return returnCode;
+
+	std::vector<BatchElement> batchTail;
+	for (int i = returnCode + 1; i < currentBatchSz; ++i)
+		batchTail.push_back(currentBatch.at(i));
+	currentBatch = batchTail;
+
+	return returnCode;
+}
+
+static int fillVideoFrameBatch(
+	MediaSources &mediaInputStruct,
+	const DataProcessingConditions &dataProcessingConditions,
+	std::vector<BatchElement> &currentBatch
+) {
+	std::clock_t start = clock();
+	int currentFrameBatchSize = 0;
+	Mat nextFrame;
+
+	std::vector<KeyPoint> nextFeatures;
+	int skippedFrames = 0, skippedFramesForFirstFound = 0;
+	logStreams.mainReportStream << "Features count in frames added to batch: ";
+	while (
+		currentBatch.size() < dataProcessingConditions.frameBatchSize
+		&& getNextFrame(mediaInputStruct, nextFrame)
+		) {
+		// TODO: here can be undistortion
+		fastExtractor(nextFrame, nextFeatures,
+			dataProcessingConditions.featureExtractingThreshold);
+		if (nextFeatures.size() < dataProcessingConditions.requiredExtractedPointsCount) {
+			if (currentFrameBatchSize == 0)
+				skippedFramesForFirstFound++;
+			skippedFrames++;
+			continue;
+		}
+		logStreams.mainReportStream << nextFeatures.size() << " ";
+		currentBatch.push_back({
+			nextFrame.clone(),
+			nextFeatures
+		});
+		currentFrameBatchSize++;
+	}
+	logStreams.mainReportStream << std::endl << "Skipped for first: " << skippedFramesForFirstFound << std::endl;
+	logStreams.mainReportStream << "Skipped frames while constructing batch: " << skippedFrames << std::endl;
+	logStreams.mainReportStream << "Batch size: " << currentBatch.size() << std::endl;
+	std::cout << "Extracting time: " << (clock() - start) / CLOCKS_PER_SEC << std::endl;
+	return skippedFrames;
+}
+
+
+static int findGoodFrameFromBatch(
+	const DataProcessingConditions &dataProcessingConditions,
+	std::vector<BatchElement> &currentBatch,
+	std::vector<bool> &isMatchesEstimated,
+	std::vector<std::vector<DMatch>> &estimatedMatches,
+	Mat &previousFrame, Mat &newGoodFrame,
+	std::vector<KeyPoint> &previousFeatures,
+	std::vector<KeyPoint> &newFeatures,
+	std::vector<DMatch> &matches
+) {
+	int batchSz = currentBatch.size();
+	Mat goodFrame;
+	std::vector<KeyPoint> goodFeatures;
+	std::vector<DMatch> goodMatches;
+	int goodIndex = FRAME_NOT_FOUND;
+	Mat candidateFrame;
+	std::vector<KeyPoint> candidateFrameFeatures;
+	std::vector<DMatch> candidateMatches;
+	for (
+		int batchIndex = batchSz - 1;
+		batchIndex >= dataProcessingConditions.skipFramesFromBatchHead;
+		batchIndex--
+		) {
+		while (!isMatchesEstimated.at(batchIndex)); // Dmitry Valentinovich, I'm sorry...
+		candidateFrame = currentBatch.at(batchIndex).frame.clone();
+		candidateFrameFeatures = currentBatch.at(batchIndex).features;
+		candidateMatches = estimatedMatches.at(batchIndex);
+
+		logStreams.mainReportStream << "Batch index: " << batchIndex
+									<< "; curr. extracted: " << candidateFrameFeatures.size()
+									<< "; matched " << candidateMatches.size() << std::endl;
+
+		if (
+			candidateMatches.size() >= dataProcessingConditions.requiredMatchedPointsCount
+			&& candidateMatches.size() >= goodMatches.size()
+			) {
+			logStreams.mainReportStream << "Frame " << batchIndex << " is a good" << std::endl;
+			goodIndex = batchIndex;
+			goodFrame = candidateFrame.clone();
+			goodFeatures = candidateFrameFeatures;
+			goodMatches = candidateMatches;
+			if (dataProcessingConditions.useFirstFitInBath)
+				break;
+		}
+	}
+	if (goodIndex != FRAME_NOT_FOUND) {
+		newGoodFrame = goodFrame.clone();
+		newFeatures = goodFeatures;
+		matches = goodMatches;
+		return goodIndex;
+	}
+
+	return goodIndex;
+}
